[Transpose] Refactor kernel functions (#546)
* Kernel header

* Implementations

* Kernel files

* Format

* Rename kernel files

* Fix transpose.h and add global

* Format
