+++ include/flexflow/ops/transpose.h
@@ -6,13 +6,6 @@
 
 namespace FlexFlow {
 
-class TransposeMeta : public OpMeta {
-public:
-  TransposeMeta(FFHandler handler) : OpMeta(handler){};
-  int num_dim;
-  int perm[MAX_TENSOR_DIM];
-};
-
 class Transpose : public Op {
 public:
   using Params = TransposeParams;
@@ -52,28 +45,6 @@ public:
                             std::vector<Legion::PhysicalRegion> const &regions,
                             Legion::Context ctx,
                             Legion::Runtime *runtime);
-  static void forward_kernel(TransposeMeta const *m,
-                             float const *input_ptr,
-                             float *output_ptr,
-                             Legion::Domain in_domain,
-                             Legion::Domain out_domain,
-                             ffStream_t stream);
-  static void forward_kernel_wrapper(TransposeMeta const *m,
-                                     float const *input_ptr,
-                                     float *output_ptr,
-                                     Legion::Domain in_domain,
-                                     Legion::Domain out_domain);
-  static void backward_kernel(TransposeMeta const *m,
-                              float *input_grad_ptr,
-                              float const *output_grad_ptr,
-                              Legion::Domain in_grad_domain,
-                              Legion::Domain out_grad_domain,
-                              ffStream_t stream);
-  static void backward_kernel_wrapper(TransposeMeta const *m,
-                                      float *input_grad_ptr,
-                                      float const *output_grad_ptr,
-                                      Legion::Domain in_grad_domain,
-                                      Legion::Domain out_grad_domain);
   bool measure_operator_cost(Simulator *sim,
                              MachineView const &pc,
                              CostMetrics &cost_metrics) const override;
